output averaged info

+++ include/thread.h
@@ -39,7 +39,7 @@
 	extern thread_tt thread_create(int);
 	extern void thread_destroy(thread_tt);
 	extern int thread_gettid(const_thread_tt);
-	extern int thread_wtotal(const_thread_tt);
+	extern double thread_wtotal(const_thread_tt);
 	extern int thread_assign(thread_tt, int);
 	/**@}*/
